Moved GridItem & Border into main components dir
